Cleanup modules
I was using bad practice in the code structure with modules so I
followed clippy's suggestions.
